Fixed both false positives and communication between client and server. Added feedback and results to client and server
Fixed both false positives and communication between client and server. Added feedback and result to the server.cpp file so as to write the results in the message readable by the client and wrote the corresponding part that allows the client to read the final result and feedback messages. Furthermore, the possibility of deciding whether to see the feedback during the robot's movement or to only see the result has been added to the client

Co-Authored-By: AndreaFel <and.felline@gmail.com>